feat: split up vision code
