+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/api/vision/VisionAPI.kt
@@ -1,0 +1,7 @@
+package org.firstinspires.ftc.teamcode.api.vision
+
+import org.firstinspires.ftc.vision.VisionProcessor
+
+interface VisionAPI {
+    val processor: VisionProcessor
+}
