+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmode/teleop/DemoTeleOp.kt
@@ -3,6 +3,8 @@ package org.firstinspires.ftc.teamcode.opmode.teleop
 import com.qualcomm.robotcore.eventloop.opmode.OpMode
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp
 import org.firstinspires.ftc.teamcode.api.DemoQuadWheels
+import org.firstinspires.ftc.teamcode.api.vision.AprilVision
+import org.firstinspires.ftc.teamcode.api.vision.Vision
 import org.firstinspires.ftc.teamcode.components.DemoForward
 
 @TeleOp(name = "Demo")
@@ -11,6 +13,12 @@ class DemoTeleOp : OpMode() {
     override fun init() {
         // Initializes an API, shared code used by components.
         DemoQuadWheels.init(this)
+
+        // Initialize the April Vision API
+        AprilVision.init(this)
+
+        // Initialize the Computer Vision API and register the April Tag processor
+        Vision.init(this, AprilVision)
     }
 
     // Run repeatedly while the robot is running.
